Add a mutex to AllocationLocationTracker functions
Summary:
In concurrent mode Hades needs to have some extra synchronization
when enabling/disabling the memory profiler, and when adding new allocs.

Add a `Mutex` used to protect the `stackMap_` member. The `enabled_` member
is protected by ensuring that no background GC is running while enabling/disabling.

The cost of acquiring mutexes is only paid when the profiler is on, which is only in
debug modes. This should have no effect on release performance.

Reviewed By: neildhar

Differential Revision: D25284428

fbshipit-source-id: 0964c070f2ad2cbe01cf004d5c5e1b9d9c814f61