Lexer stores previous token's end source location
Summary:
To support upcoming changes to source locations for parenthesized types and expressions, we must maintain the end source location for the previous token. This will allow us to find the source location of the last right paren enclosing a type/expression, after the lexer has moved on to the next token after the right paren.

The lexer now contains a new field, `prevTokenEndLoc_`, which can be accessed from either the lexer or parser via the new `getPrevTokenEndLoc` method.

The `prevTokenEndLoc_` field is only set when a new token is produced. I opted to create a small utility method `finishToken` which is called right before returning from the three lexer methods which return a new token: (`advance`, `advanceInJSXChild`, and `rescanRBraceInTemplateLiteral`). The `finishToken` method takes the new end location for the token, saves the old end location in `prevTokenEndLoc_`, and updates the current token's end location to the new end location.

This means that we must avoid setting the end location for the current token until `finishToken` is called right before returning (otherwise we would clobber the last end location which should be saved). Luckily we can just remove any `token_.setEnd` calls as the token's end location will be set by `finishToken`. The only caveat is that the errors in `scanNumber` now need to use the current char ptr as the end location instead of first storing it as the token's end location, but this is straightforward as we can just switch to calling `errorRange`.

Also note that we must be sure to save and restore the value of `prevTokenEndLoc_` in lexer save points.

This diff only introduces the changes to the lexer - the diffs stacked on top will use `getPrevTokenEndLoc` for the new parens parsing scheme.

Reviewed By: tmikov

Differential Revision: D25291364

fbshipit-source-id: 8a1104bd51e200224cd3c361d65928ef1e6e4789