Dirty cards and evacuate compactee during YG collections
Summary:
Slightly update the criterion for evacuation so that compactee objects
will now be evacuated by `EvacAcceptor`.

In addition, an object promoted from YG to OG in the middle of an OG
marking will not get scanned by the OG, because it will be promoted
as black and will not get marked. The SATB mechanism means that
this is not a problem for marking. However, for accurate compaction,
we need to scan the promoted object for pointers into the compactee,
and appropriately dirty cards for its new location.

Reviewed By: dulinriley

Differential Revision: D24388210

fbshipit-source-id: 6506aa73ebe40b58003f5ac047d2f8ee7bb93768