Dirty cards in write barriers
Summary:
Dirty cards for pointers into the compactee in the existing
generational write barrier. In order to avoid erasing card table
information for a pending compaction, do not erase card tables when a
compaction is pending.

During compaction, skip scanning cards for the compactee segment
since it is going to be fully traversed anyway.

Reviewed By: dulinriley

Differential Revision: D24386071

fbshipit-source-id: 93881a54dd28ef08c3333cc66507260b3243f23c