Fix up stats for compaction
Summary:
Add compaction information to the YG collection stats. I considered
two other ways of doing this:
1. Adding compaction information to the OG stats: this would require
some changes to how we handle OG stats, and will delay things like
the tripwire check when a compaction is happening. Juggling the extra
stats values seems bug prone and it seemed significantly simpler to
just edit the existing YG stats.
2. Adding compaction stats as a separate stats entry: This would be
fine, but it would require extra special casing and all the reported times
would be identical to the YG collection anyway. In addition, we would
need a way to distinguish bytes evacuated from the YG vs bytes
evacuated from the compactee. Finally, this involves an extra call to
the analytics callback.

Overall, since the most interesting stats we log are the times, and
because compactions are fairly rare compared to normal YGs, I decided
to just do a special version of the YG stats logging, which
distinguishes it from non-compacting YGs.

I'm open to revising this if there are strong objections.

Reviewed By: dulinriley

Differential Revision: D24472340

fbshipit-source-id: a044da43d7707ad4bff4322d435230f195a0c863