Add ability to remove OG segments
Summary:
Add the ability to remove segments from the OG. This is a very
expensive operation since it has to make invasive changes to a number
of data structures. We could make this cheaper by adapting the data
structures to allow unoccupied slots in them, but that would require
broader changes to the code and several additional checks.

Reviewed By: dulinriley

Differential Revision: D24715308

fbshipit-source-id: 647833b2a6f6358222a90b14007507b2c23fd64d