Add native memory tracking to JSWeakMapImplBase
Summary:
The JS objects `WeakMap` and `WeakSet` anchor some native memory with
their `map` members that should be reported. Add this native memory via some
`_snapshotAdd*` callbacks.

Normally, native memory is tracked using its address, so it only takes space if
snapshotting is turned on. Unfortunately, when using a DenseMap, we don't know
the internal address it uses and when it changes, so we can't use that same mechanism.

Instead, add a `IDTracker::getExtraNativeIDs` function that can associate additional
native memory with an object while it's alive. The extra IDs are uniquely owned by
the heap memory.

Reviewed By: neildhar

Differential Revision: D25597805

fbshipit-source-id: 86a0bbd75d5e63a08084050363b0b79332034e54