Use DummyRuntime::getHeap in unit tests
Summary:
Quite a few unit tests relied on the `gc` member of `DummyRuntime`
being public and accessing it directly. This will complicate things if
we change the type of the `gc` member, so make an accessor instead
called `getHeap()` that mimics the normal Runtime's accessor.

For any tests that rely on a specific GC, do a `static_cast` to the more
narrow type, and use any public static constants directly on the more
narrow subclass of GCBase.

Also refactor some of DummyRuntime to call `getHeap()` itself.
These are just used in tests, so there should be no effect on
runtime performance or correctness.

Reviewed By: neildhar

Differential Revision: D25505874

fbshipit-source-id: 9f521c4233897e24f8d506b5ea253be309b75af8