Runtime::heap_ -> Runtime::getHeap()
Summary:
The type of `heap_` can change in the future, and there's already
an accessor for it called `getHeap()`. Use that throughout
Runtime.cpp to reduce the amount of changes associated with
changing the type of `heap_` in the future.

This should have no performance or correctness impact, as `getHeap()`
is a one-liner in the header file.

Reviewed By: willholen

Differential Revision: D25510489

fbshipit-source-id: 0a19d94a1466bde8eb10756839acb78f5e9484fa