Add Compact mode to ESTreeDumpJSON.
Summary:
Add a new mode to ESTreeDumpJSON which ignores any empty nodes,
unconditionally. This will result in faster dumping of smaller ASTs,
as there are no string map lookups required. The tradeoff is that the
consumer of the dump must account for the missing properties using
some kind of defaulting mechanism.

Differential Revision: D25037552

fbshipit-source-id: b6b510f521b999cc7aba5bff83ee46cdcab4fc43