Shrink the HeapSnapshot::NodeID to take less space
Summary:
The HeapSnapshot::NodeID doesn't need a full 2 ^ 64 ids.
The number of IDs that exist is going to be roughly proportional
to the size of the heap, which can't go above 4 GB in a single instance.
It's also a significant factor reduced from the size of the heap, since it's one
ID for every object, not for every byte.

In practice, there should never be close to enough objects to reach 4 billion total
ids used. This can save space in anything storing IDs.

Reviewed By: neildhar

Differential Revision: D25599579

fbshipit-source-id: 8b09d9cf9683e294188ebc818acea715fabf938f