Use compressed pointer indices for CrashManager names
Summary:
Update the CrashManager naming scheme for Hades to follow the
compressed pointer index of a given segment. This is in addition to
entries tagged "YG" or "COMPACT". This means that a YG or compactee
segment will have duplicate entries in the crash manager, which
should make it easier to debug (and simplifies the code).

Reviewed By: dulinriley

Differential Revision: D25662167

fbshipit-source-id: f717500ec87b6c3536dc3a061aeb871f3f4aebce