Set up initial hermes-eslint library
Summary:
The next step for the WASM Hermes parser is adding support for it to be used as a custom parser for ESLint. Let's start setting up the `hermes-parser` library. The `hermes-parser` library should export a single function, `parseForESLint`, that returns an object containing the ESTree AST and optionally contains visitor keys and a scope manager (which will be added in later diffs) (https://eslint.org/docs/developer-guide/working-with-custom-parsers).

Let's rearrange our source files slightly by moving all the WASM parser files under the new `src/hermes-parser` directory and all new ESLint files under the new `src/hermes-eslint` directory. Let's then update the tests and build scripts to account for this.

ESLint requires an array of tokens on the Program node, so let's enable the new `tokens` option in `hermes-eslint`.

Reviewed By: gkz

Differential Revision: D25399672

fbshipit-source-id: ef041268b2c3483333b53d7c8c1776fa119228bd