Optionally include tokens in parser output
Summary:
Now that tokens are stored in the lexer, lets update the WASM parser to optionally return a list of all tokens on the Program node. This can be enabled with the `{tokens: true}` option, just like in Babel. These tokens will be needed for the upcoming use of the Hermes parser for ESLint.

ESLint and other consumers of tokens expect a general token type for each token. We can base this off of esprima's token types (which Babel also uses), which includes types such as "Identifier", "Keyword", "Punctuator", etc. We must then categorize each token into one of these types.

Serializing tokens to send them from WASM to JS is done in the same way as comments. We first write the token type as an integer, then write the source location, then use the source location to write the token's value as a string (which consists of a pointer followed by size). When deserializing we maintain an array of token type names in JS that we can index into using the serialized token type integer.

Note that we do not need to include the EOF token since this is represented by the end of the list of tokens (consumers such as ESLint also do not expect an explicit EOF token).

Reviewed By: avp

Differential Revision: D25617465

fbshipit-source-id: 8540679d275cc0e20a411f20c39f4c8a9574fda1