Do not count sweeping towards drain rate
Summary:
Counting sweeping against the number of iterations we have until the
OG is full causes us to overestimate the amount of OG marking work we
really need to do in order to keep up with demand. Instead, just
assume that we will free space quickly enough as we sweep.

We could improve on this using lazy sweeping, so the allocation path
sweeps just enough to be able to proceed with an allocation, in cases
where the heap is very full.

Reviewed By: dulinriley

Differential Revision: D25689005

fbshipit-source-id: 5c658270c0f9e32b0877f9e57d267bfda3937566