Make PinnedSymbolID
Summary:
Create a new PinnedSymbolID to be used for roots. This has 3 main
advantages:
1. We can fully distinguish between roots and heap values in the GC.
2. We can separate the RootAcceptor and SlotAcceptors completely.
This would also serve as an enforcement mechanism, since
RootAcceptors would only be able to accept PinnedSymbolIDs and
SlotAcceptors will only be able to accept GCSymbolIDs.
3. Heap and root values theoretically have different concurrency
requirements, since roots are marked in a STW pause. This could be
useful in future.

Reviewed By: dulinriley

Differential Revision: D25746215

fbshipit-source-id: 57f2528ccade6c974c77b60a59de8e221ff7dceb