Assert no ongoing collection in Hades destructor
Summary:
There should be no remaining collection in the background thread
during the Hades destructor, since `finalizeAll` calls
`waitForCollectionToFinish`. We should assert that
`concurrentPhase_==Phase::None` to ensure that
`waitForCollectionToFinish` was called correctly.

This is to avoid a failure mode where the destructor deadlocks
because the background thread is waiting for an STW pause.

Reviewed By: dulinriley

Differential Revision: D25748321

fbshipit-source-id: 8ab3a1ab576f8502b18de691d7dee66c93d11bfc