Trim cells during compaction in Hades
Summary:
Some cells that have excess capacity can be trimmed when they're moved.
Compaction is a great time to trim off the excess capacity on arrays and property
storage. This can save a lot of memory if those things aren't resized again.

Do *not* do this for objects promoted from YG, because that could cause it to
cycle back and forth between YG and OG if it gets resized. Only do it for objects
in OG being compacted, since those are less likely to change in size.

Should be a noticeable (up to 10%) memory win on benchmarks.

Reviewed By: neildhar

Differential Revision: D25765075

fbshipit-source-id: 1683234520eb52f5c92d8ef40b5ab908ceedabfa