Change getTrimmedSize() to assert that it can be trimmed
Summary:
The previous interface for `getTrimmedSize()` returned its second argument if
the cell didn't allow for trimming. This was meant to be a convenience thing,
but most of the callers had to know whether the return value was the same for other
reasons anyway, so the convenience wasn't needed.
Change the interface to assert that trimming is enabled when `getTrimmedSize()` is called.

Reviewed By: neildhar

Differential Revision: D25765076

fbshipit-source-id: 288de6c3777df9722e2b0640c886db31d21c7007