Remove increase/decrease domain count
Summary:
Stop tracking the exact domain count in the sampling profiler. This
approach is fragile since it relies on each domain incrementing a
count with the sampling profiler.

Instead, we can avoid allocations in the signal handler by ensuring
that we have enough space for the maximum possible domain count that
may be added inside the signal handler (which is equal to the maximum
possible stack depth).

In addition, we should clear all profiler data when the runtime is
unregistered, to avoid accessing the freed memory.

Reviewed By: dulinriley

Differential Revision: D25718125

fbshipit-source-id: 4bf4323bf15165fa512ea8893cc2d4105098d1c3