Add a GCScope to createHermesInternalObject
Summary:
Creating the HermesInternal object now makes a lot of handles,
so it needs its own GCScope.
The global object GCScope is also no longer large enough when
Handle-SAN is turned on with GenGC, so bump the limit slightly.

Add a test to ensure that a basic runtime can be constructed for an empty
JS file even when Handle-SAN is turned up to the max.

Reviewed By: Huxpro

Differential Revision: D25796305

fbshipit-source-id: 9e30a4f671c42f7041e54221aa3dfdacfc9b4134