Add size change updates to GCBase::moveObject
Summary:
There was a bug in how `AllocationLocationTracker` handled objects that could
be trimmed: it would miss decrementing the live byte count by the trimmed amount.
Add the possibly changed sizes to `GCBase::moveObject`.
This will also be used for any GCs that can increase the size of an object over time,
for example, moving from YG to OG might change an object's size in the future.

Unfortunately in order to access the sizes in GenGC it needs to be moved from a separate
code path into the sweeping loop. That's the only place it knows both the original sizes
and the trimmed sizes.

Reviewed By: neildhar

Differential Revision: D25789121

fbshipit-source-id: c5c2fdcf38eb830118cf520a1c1271fe0d53674b