Add a cleanup phase to OG collections
Summary:
We currently have some special handling in `oldGenCollectionWorker`
to ensure that the mutator is not able to observe a collection has
ended and reset the stats for it. This requires that we perform the
stats manipulation under a lock right after setting
`concurrentPhase_` to `None`. Instead, add a new `Cleanup` phase to
manage the final stats update. This phase is only used in concurrent
collections.

Reviewed By: dulinriley

Differential Revision: D25690008

fbshipit-source-id: 420a6222840abfc470b1494eb84903fdccf9aa54