Perform blocking collections on the mutator
Summary:
We currently have an issue in with blocking concurrent collections as
they do not count towards the CPU time used by the mutator. This diff
simplifies blocking collections so that they are performed entirely
on the mutator, which allows us to keep track of the CPU time spent
in them.

More generally, this allows us to mix concurrent and incremental
collection modes, which could also prove useful for other things in
future.

Reviewed By: dulinriley

Differential Revision: D25683637

fbshipit-source-id: 5137f5c99f7e509e1287b47a4259e671211e7866