Do not release lock during signal handler
Summary:
We should hold the profilerLock for the entire duration of
`sampleStack`, including during the signal handler. This guarantees
that the mutator does not hold the profilerLock when the signal
handler is triggered and therefore ensures that the profiler state
accessed by the signal handler is not corrupted. Whether this is
actually possible depends on how pthread_kill is synchronised with
the mutator (specifically, whether the mutator can acquire
profilerLock before observing the signal).

Either way I don't think there's an issue with sending a signal to
the mutator while it is waiting for the profilerLock (as is currently
described in the code) because a thread can still handle signals
while waiting for a mutex.

Reviewed By: dulinriley

Differential Revision: D25740075

fbshipit-source-id: 9490dba5aefdd21bd24187f744becf4cd0af95b2