EASY: use CAPTURE_IP_ASSIGN only when needed
Summary:
The difference between CAPTURE_IP() and CAPTURE_IP_ASSIGN() is subtle
and can create confusion. CAPTURE_IP_ASSIGN() should only be used when
we need to declare a new variable and assign a result to it.

Enforce this rule. This improves readability because many assignments
become explicit and the remaining ones have a declaration syntax, which
is sufficient hint of what is going on.

Reviewed By: mhorowitz

Differential Revision: D23675350

fbshipit-source-id: 4013bb77fe47f672002c60bc1ded39153a6c1b49