Enforce memory ordering with signal handler
Summary:
TSAN currently complains that we do not ensure that access to
domains_ and sampleStorage_ is synchronised. Since we cannot actually
acquire the profilerLock in the signal handler, we can perform this
synchronisation with a sort of hand rolled atomic spin lock.

The atomic store flushes all writes from one thread to another, and
the other thread waits until it has observed the store before
proceeding.

Reviewed By: dulinriley

Differential Revision: D25841614

fbshipit-source-id: 271349328cee02435ec17f1e0198e665d810b8e9