Skip trimming for correctly sized cells
Summary:
If a cell is already the right size, there is no need to trim it
again. It is slightly cleaner to express this without the
`canBeTrimmed` function, so this diff removes that as well.

This change may be a marginal perf bump, and will have
no impact on memory consumption.

Reviewed By: dulinriley

Differential Revision: D25880048

fbshipit-source-id: f13d576813cfe8c7fe406edffe2a2f3e47385fc3