re-enable TDZDedup for the new TDZ scheme
Summary:
TDZDedup changes a little because it now must track the possibility that
a value transitions from back available to unavailable when "empty" is
stored into it.

Additionally, TDZDedup now relies on TypeInference because it needs to
track the type of the value being stored.

It is still likely less effective than before. One example is assigning
the return value of a call to a TDZ variable. If we don't have type
inference information of the function call result (which is most of the
type), we have to assume that it may become "empty" again. We may want
to assume that functions never return "empty", which is true in
practice right now, but such an assumption would be risky, because it is
possible to construct valid IR where "empty" is returned from a
function.

Reviewed By: avp

Differential Revision: D25750408

fbshipit-source-id: f3cc1bd379760ea67636459fa9b4f1c49c7ee2eb