Separate local state of sampling profiler
Summary:
There are currently a few issues with supporting multiple runtimes in
the sampling profiler. For instance:
1. There is only a single global list of domains that need to be kept
alive for symbolication, which each runtime marks as roots. To
support multiple runtimes, each runtime should have its own
associated list of domains.
2. The sampled stacks across all runtimes are stored in a single
list, which makes it difficult to pick apart the profiling data for
each runtime.

This diff separates the state of the sampling profiler into two
classes:

## `SamplingProfiler::GlobalState`
This singleton class maintains the global state of the sampling
profiler. It maintains signal handler registration, whether the
profiler is enabled/disabled, and the lifetime of the timer thread.
Each runtime must be registered with this global object. The timer
thread will iterate over all registered runtimes and sample stacks
for each of them.

## `SamplingProfiler`
This class acts as a guard to be stored in each runtime, and
maintains all of the profiler data for that runtime.
Creation/destruction of this class will register/deregister the
associated runtime from the global profiler on the current thread.
Captured stacks and the domains we need to keep alive are stored in
this class, since they are only associated with a specific runtime.

With this implementation, the `SamplingProfiler` object needs to be
recreated whenever the runtime is moved onto a different thread. We
may want to revisit that in future, but it is acceptable for now.

Reviewed By: dulinriley

Differential Revision: D25734463

fbshipit-source-id: cc4c7e2b992279805a959a25ab4239604ac1550e