Parser option to set sourceType
Summary:
We currently set the `sourceType` on the program node to "module" if any ES6 imports or exports are detected, otherwise we set it to "script". We should also allow the option to specify the `sourceType` in the options passed to the parser, just like Babel does.

Let's add a `sourceType` option which can be either "module", "script", or "unset". This value will be used as the Program node's `sourceType` if set, otherwise the Program node's `sourceType` will be autodetected using the same logic as before.

Reviewed By: gkz

Differential Revision: D25981476

fbshipit-source-id: e663d67b24322656b05b1d3a7b8a2c37afd727b2