Remove unneeded scope manager properties
Summary:
In preparation for augmenting the scope manager for handling types let's clean up the existing scope manager code a bit.

We can remove a few properties that are not needed by ESLint (https://eslint.org/docs/developer-guide/scope-manager-interface). In particular we can remove the `tainted` and `partial` properties of `References`, the `taints` property of `Scopes`, and the `stack` property of `Variables`. Let's also rename the `__left` property of `Scopes` to be `__referencesLeftToResolve` for clarity.

Reviewed By: gkz

Differential Revision: D26022319

fbshipit-source-id: b67e435fc6ce28fe7ad9c9e751f277b7f5dc8989