Create initial Hermes scope manager with options set
Summary:
Now that we have pulled in a forked version of `eslint-scope`, lets start to create our custom scope manager for `hermes-eslint`. This diff creates the initial `HermesScopeManager` using the `eslint-scope` utilities, and generates a scope manager in the top level `parseForESLint` function.

`eslint-scope` has a number of options that be configured, however for the purposes of Hermes's scope manager almost all of these can be hardcoded. In fact let's just inline the hardcoded values of these options through the source code to clean it up a bit. The options that we can effectively inline through the rest of the scope manager are:
```
{
  ignoreEval: true,
  optimistic: false,
  directive: false,
  nodejsScope: false,
  impliedStrict: false,
  ecmaVersion: 6,
}
```

These are generally the same options used by `babel-eslint`, and we can treat all parsed code as ES6+ as we are parsing ES6+JSX+Flow code.

The one option we need to keep is the `sourceType` if it is specified. This will be passed as a property on the optional options parameter (the second parameter for `parse` and `parseForESLint`). We can just forward the `sourceType` property to the Hermes parser, then read from the `sourceType` property on the resulting AST.

We also still need to pass the `childVisitorKeys` as part of the options object for the scope manager, as the visitor keys are passed to `esrecurse` for general AST traversal.

Reviewed By: gkz

Differential Revision: D25982036

fbshipit-source-id: 2783916c723ebcdbabc430b1c99d8feccd19fc9e