Take GCCell*& in acceptors
Summary:
Instead of accepting a `void*&` in our acceptors and then casting it
to a `GCCell*&`, it is clearer to directly take a `GCCell&*` from the
start, since we can only accept pointers to GCCells anyway. To
facilitate this, update the API for `GCPointer` to reflect that it
can also only store GCCell pointers.

This is still not technically type safe, and we still need to use
`reinterpret_cast` in `acceptPtr`, because a subtype of GCCell, like
`StringPrimitive*&` cannot actually be safely converted to a
`GCCell*&`.

Reviewed By: dulinriley

Differential Revision: D26035289

fbshipit-source-id: 61ca2722b0cf05035dae46e787e4c1f4cccd739f