Fix heap timeline issue with emitting source locations more than once
Summary:
There was a bug in heap timelines that would show an object's allocated
stack as having functions calling themselves recursively, where no recursion
was possible.

The cause was a serialization error, where we were outputting the `trace_function_infos`
for a given line/col pair several times with the same ID. When Chrome visualized the
timeline, it would accidentally replace other valid parts of the stack with the duplicated
function source location.

There was also a small simplification in the same area, where the `SourceLocMapInfo`
struct members were duplicated (for no reason I could see) in the body of
`emitAllocationTraceInfo` instead of re-using the existing map info.

Included with this diff are some changes to `chrome_snapshot_parser.py` that helped
me identify this issue: don't delete the `trace_function_infos` part of the file, and don't
add the +1 to lines/columns that are already one-based.

Reviewed By: motiz88

Differential Revision: D26065806

fbshipit-source-id: fbfef7a23c721dbfb2b00e638ccf2823af2c4382