Make write barrier APIs consistent across GCs
Summary:
To remove some ifdefs from GCHermesValue and GCPointer, make the
same APIs available on all GCs. For GenGC, all the snapshot write barriers
are no-ops, and for MallocGC all forms of barriers are no-ops.

Define them as no-ops in GCBase for convenience, and use a static overload
for Hades to do something different. In some cases, improve the code from
calling a generic write barrier to `snapshotWriteBarrier` instead.

Reviewed By: neildhar

Differential Revision: D25985042

fbshipit-source-id: 77b6efaf88fb73a66161915aabdbe22ef438ed31