Make GCBase methods virtual
Summary:
Make all the functions that need to be provided by GCBase subclasses
virtual, to both enforce that the type signatures match, and to allow using
a `GCBase *` in some places instead of `GC *` where performance isn't
critical.

Since all GC subclasses are `final`, and the typedef for GC to a final subclass means
that all of the virtualization will be devirtualized at compile time, and there should be
no runtime cost to this change.

Currently this diff doesn't change anything to actually take GCBase, because it isn't
necessary. But some debugging APIs could easily be changed if necessary.
Also, `GCBase::createSnapshot` *could* be refactored after this diff,
but it isn't necessary right now.

Reviewed By: neildhar

Differential Revision: D25470951

fbshipit-source-id: cc681733eba5d4ff11d13649017bf41793600a02