Fix CJS source-maps of shared-modules
Summary:
Hermes generated incorrect source-maps for segments containing a shared module (a module that is part of more than one segment). The issue was that modules are not correctly de-duplicated before passing them to the `SourceCodeGenerator`. The issue only manifests if:

1. it isn't the first segment containing the shared module
2. the shared module comes before the module referenced by the stack frame.

The existing test didn't catch the error because (2) wasn't given.

This diff ensures that modules get correctly de-duplicated.

Reviewed By: motiz88

Differential Revision: D26127122

fbshipit-source-id: 865aa491bf8375459ca9d190532e399557bd7339