Split definitions into subclasses with type information for references and definitions
Summary:
Let's augment Definitions and References with whether they are for types or values, in preparation for adding Flow-specific scope information.

References can either be a type reference or value reference, so we only need a single boolean flag denoting this.

Definitions can introduce either a value, type, or both (e.g. class and enum definitions are both types and values). Instead of directly creating `Definitions`, let's create `Definition` subclasses for each type of definition, just like the one that already exists for `ParameterDefinition`, and exclusively use these instead. This makes it clearer how each definition is constructed, and allows us to set whether each type of definition is a value definition, type definition, or both. We also move the types of definitions to the `DefinitionType` object instead of using properties added to the `Variable` class, as these types are actually referring to definitions not variables within the scope manager.

We have not yet introduced any pure type definitions yet, but these will be implemented soon (e.g. TypeAliases and TypeParameters). There are three types of definitions currently added that can be both type and value definitions - class names, implicit global variables (which may be implicit global values or types), and import bindings (as value imported bindings may be used as types, e.g. for imported classes. We can restrict pure type imports to be pure type definitions in the future once we check for type imports).

Finally we can introduce a utility for creating type references vs value references, and check that type references match to type definitions and value references match to value references when resolving names. This will be used in future diffs where we implement Flow-specific scope information.

Reviewed By: gkz

Differential Revision: D26030813

fbshipit-source-id: 478fc4cac14e537315628078f417c2ea3923e284