Prevent incorrect identifier references in types
Summary:
Whenever the scope manager's referencer visits an Identifier node, it treats it as a value reference. However Flow type nodes introduce a few places where Identifier nodes should not be treated as references, as they are instead property or parameter names. In all of these cases all we need to do is prevent visiting the Identifier node, though we still want to visit all other children of the parent node.

There are 5 cases of this we currently need to prevent. These same cases are handled in `babel-eslint` by completely deleting the offending Identifier nodes from the AST, but thankfully with a custom scope manager we can prevent this and simply avoid visiting them so they will not be treated as references.

- `QualifiedTypeIdentifier`: Only the first component of the qualified type identifier is a reference (and it is actually a value reference). Here we need to walk up the `qualifications` until we find an `Identifier`, and ignore the `id` child nodes as these are all properties.
- `ObjectTypeProperty`: The `key` child may be an identifier like 'foo' in `type t = { foo: string }`. We can simply prevent visiting the `key` child if it is an Identifier, since it is actually a property name.
- `ObjectTypeIndexer`: If the `id` child exists it is always an unreferenceable parameter name that we should avoid visiting (e.g. 'foo' in `type T = { [foo: string]: number }`)
- `ObjectTypeInternalSlot`: The `id` child is the internal slot's name, not an identifier that can be referenced (e.g. 'foo' in `type T = { [[foo]]: number }`)
- `FunctionTypeParam`: The `name` child only exists for readability, and is not actually the definition or a reference of a parameter name. (e.g. 'foo' in `type T = (foo: string) => void`)

Reviewed By: gkz

Differential Revision: D26057884

fbshipit-source-id: 37f416c531861d1c42f689f9c52d7a690d39e1d0