Default to module sourceType in scope manager
Summary:
The scope manager allows for setting the sourceType to either "script" or "module". If a sourceType option is not provided the scope manager currently attempts to infer the sourceType value from the AST itself based on the presence of ES6 imports/exports. Instead, we should be defaulting to the "module" option if no sourceType was explicitly provided. This is done both for compatibility with `babel-eslint` which defaults to a "module" sourceType as well, and because the "module" sourceType has cleaner scoping rules overall and is usually what users would want when analyzing a single file unless they explicitly opt in to "script" mode.

This is a simple change to make but we must take care to not break existing tests. All the tests forked from `eslint-scope` assume that the default sourceType is "script", so we can create a new `eslint-scope-test-utils` file that provides a `parseForESLint` function that defaults sourceType to "script" for just these tests. We then just point to using this testing function for all tests forked from `eslint-scope` so that they will continue to work as before.

Reviewed By: gkz

Differential Revision: D26132037

fbshipit-source-id: f8246f97868e607039b5e84ea1c486653da7b0c1