Store strictness information during preparse.
Summary:
Preparsing properly parses directives when running on functions,
and stores strictness information in the functions' AST node decorations.
That information was not shared with the lazy parse step, which only
stored start/end locations for functions.

Augment the preparsing data to also store strictness information for
these functions and read it back in during the lazy parse.
This also requires the semantic validator to set the strict mode
based on the node strictness instead of reverifying it via scanning
the prologue of the function for directives (in lazy parse mode,
those bodies are empty).

Reviewed By: tmikov

Differential Revision: D25318752

fbshipit-source-id: 5b53f6cae31408dc034d8b0e2f30a929c95d3cc8