Visit Flow-specific properties of functions and classes in scope manager
Summary:
All Flow nodes have been implemented in the scope manager. Now there are just two locations left where a handful of Flow-specific nodes are being skipped in the original implementation of the scope manager. These two locations are when visiting classes or functions (of any declaration/expression/arrow type). When visiting functions we must be sure to also recurse into the `returnType` and `predicate` properties, and when visiting classes we must be sure to recurse into the `superTypeParameters` and `implements` properties.

With this done, all Flow nodes/properties should be correctly visited.

Reviewed By: gkz

Differential Revision: D26140602

fbshipit-source-id: d03531b100d14639e85fd146f42981265c450b7a