Handle OptionalMemberExpression properly in scope manager
Summary: The property of member expressions may be an identifier, but this identifier is a property name not a reference to a particular definition. This is already handled properly for `MemberExpressions`, but we should add the same logic for `OptionalMemberExpressions` as well. This simply requires factoring out the shared logic into a `visitMemberExpression` helper.

Reviewed By: gkz

Differential Revision: D26150706

fbshipit-source-id: 5631164ce80aed3c7c1a59b1c33ef4f8f51a1c02