Use depth tracker more aggressively in Proxy
Summary:
The code previously used a depth tracker in entry points to
Proxy to avoid infinite recursion in proxy cases when the trap was not
called.  This is insufficient, since there are some cases where an
infinite recursion can be caused in the target descriptor invariant
checks.  This diff moves the depth tracker to be used on all proxy
calls, trap or not, to avoid any other lurking bugs.  As each entry
point creates a GCScope, there's no need for a second one where the
depth tracker used to be, so those are removed.  A repro for the bug
is added to the proxy test case.

Reviewed By: avp

Differential Revision: D26142146

fbshipit-source-id: 9a54a7628ddc257d0c08257eeef677fd5db5038b