Rename isOldGenMarking_ to ogMarkingBarriers_
Summary:
With the previous diff, the value of `isOldGenMarking_` is not
directly tied to whether the OG is marking, because it remains true
only halfway through the `CompleteMarking` phase. Instead, rename it
to `ogMarkingBarriers_` which represents that the value is used to
toggle the barriers.

In line with that, change the criteria for using the
skipGarbageCallback on the compactee segment from `isOldGenMarking_`
to `compactee_.evacActive()`. This works more directly, because
evacActive can only be true once marking is fully complete, at which point
we assume parts of the OG may be swept, whereas the barriers may be
turned off before marking is done.

Reviewed By: dulinriley

Differential Revision: D26208066

fbshipit-source-id: f944f7166203ccfa97d280642ae4a388143a0597