Do more work in incremental collections
Summary:
Our incremental collections can currently be more prolonged than
necessary, because we do a fixed amount of work in each iteration. In
cases where the pause is very short, we could do more work in order
to avoid needing to grow the heap or immediately go back into a full
collection once the current one completes. This is important because
we don't actually check the occupancy of the OG while a collection is
ongoing (since we can't start another collection at the same time).

This diff adds a conservative check, such that if after the YG
collection and one step of the incremental collection, we are still
under 25ms, then it is completely safe to do another iteration.

This also reduces the probability of the following failure mode. Since
we do not budget for the iterations spent sweeping, it is possible for
the OG to be well over 70% full by the time sweeping is done. This
places additional pressure on the marking rate for the next full
collection, resulting in increasingly longer pauses, and causing more
frequent collections. With this change, we do more work proactively,
so we are more likely to complete collections early.

In future, if we find this still too conservative, we can make the check
more aggressive, to get us closer to 50ms.

The primary downside of this diff is that incremental collections will
no longer be deterministic, although we can always turn this off to get
determinism.

Reviewed By: dulinriley

Differential Revision: D25937878

fbshipit-source-id: 05251f0640b8b5a99824d27fe25085d8a5e6948b