Option to allow return outside function
Summary:
The Babel parser provides an `allowReturnOutsideFunction` option which prevents the parser from erroring on return statements at the top level. This options is always enabled for the `babel-eslint` backend, so for Hermes to replace the Babel parser as a backend for ESLint we need to support this as well.

This change adds an `allowReturnOutsideFunction` option to the parsing context which prevents the semantic validator from erroring on returns outside functions. We can then thread this value through the WASM parser, and enable it for the Hermes ESLint backend.

Reviewed By: avp

Differential Revision: D26153209

fbshipit-source-id: c764490d3bec40ac4caafbf47c3cf970faa97c29