Fix visitor key order for code path analyzer
Summary:
ESLint contains a code path analyzer that conducts some control flow analysis (https://eslint.org/docs/developer-guide/code-path-analysis). This code path analyzer requires that the children in certain control flow statements be visited in a particular order in order to work correctly.

There are a handful of visitor keys in `hermes-eslint` that are currently generated in the wrong order for ESLint's code path analyzer. Let's fix these by overriding the order of keys for these nodes in `HermesESLintVisitorKeys`. The only visitor keys we need to fix are for `IfStatements` and `ConditionalExpressions` which should be `['test', 'consequent', 'alternate']` instead of the current `['test', 'alternate', 'consequent']`, and the visitor keys for `WhileStatements` which should be `['test', 'body']` instead of the current `['body', 'test']`.

Reviewed By: gkz

Differential Revision: D26234761

fbshipit-source-id: 4525a0358bc1ef9efd29f266e4bf72a911a2c765