Add compiler options for debug levels
Summary:
Adds the `-g0`, `-g1`, `-g2` and `-g3` options to the compiler to control the *debug level*. Some of these debug levels already existed but weren't formalised / exposed in this way. In particular it wasn't possible to get the equivalent of `-g2` while emitting debug info in the bytecode.

The debug levels are:
* `-g0`: emit no debug info
* `-g1`: (the default) emit source locations of throwing instructions (e.g. for error stack traces)
* `-g2`: emit source locations of all instructions (e.g. for profiling)
* `-g3`: (also simply `-g`) emit source locations of all instructions + lexical info + additional instructions for debugger support

As before, the debug info is emitted as part of the bytecode by default, but the `-output-source-map` flag causes it to be written to a separate source map file instead. Note that `-output-source-map` implies `-g2` or higher.

Reviewed By: tmikov

Differential Revision: D26183647

fbshipit-source-id: 26414a1cc0d809854a4fc0f79142f706dbbb63c6